refactored IntoGeojson into Vertices and Edges traits
